Replace messageParser util with FilteredMessage component
This makes using the message parser easier.
